Move controller messenger type demonstration to its own test suite
